moved test_partfunc in levels
