+++ radis/levels/partfunc_cdsd.py
@@ -374,5 +374,5 @@ class PartFuncCO2_CDSDcalc(RovibParFuncCalculator):
 # %% Test
 if __name__ == '__main__':
 
-    from radis.test.lbl.test_partfunc import _run_testcases
+    from radis.test.levels.test_partfunc import _run_testcases
     print('Testing parfunc: {0}'.format(_run_testcases()))
